dashboard/requirements.txt
# Dashboard — Streamlit app (live feed, Q&A, metrics)
streamlit==1.45.1
langchain-community==0.4.1
langchain-core==1.3.2
weaviate-client==4.21.0
sentence-transformers==5.4.1
python-dotenv==1.2.2
requests==2.32.3
